requirements.txt
Pillow>=9.0.0
numpy>=1.21.0
scikit-learn>=1.0.0
Flask>=2.3.0
Werkzeug>=2.3.0
